Refactor: namespace ABFs_Construct::PCA (#3546)
