+++ source/Makefile.Objects
@@ -507,7 +507,7 @@ OBJS_MODULE_RI=conv_coulomb_pot_k.o\
       exx_abfs-jle.o \
       exx_abfs-io.o \
       exx_abfs-construct_orbs.o \
-      exx_abfs-pca.o \
+      ABFs_Construct-PCA.o \
       exx_opt_orb.o \
       exx_opt_orb-print.o \
       exx_lip.o\
