+++ source/module_ri/Matrix_Orbs11.hpp
@@ -9,7 +9,6 @@
 #include "Matrix_Orbs11.h"
 #include "RI_Util.h"
 #include "module_hamilt_pw/hamilt_pwdft/global.h"
-#include <sys/time.h>
 
 template<typename Tdata>
 RI::Tensor<Tdata> Matrix_Orbs11::cal_overlap_matrix(
@@ -130,9 +129,6 @@ std::map<size_t,std::map<size_t,std::map<size_t,std::map<size_t,RI::Tensor<Tdata
 	const ModuleBase::Element_Basis_Index::IndexLNM &index_r, 
 	const ModuleBase::Element_Basis_Index::IndexLNM &index_c ) const
 {
-// std::ofstream ofs(GlobalC::exx_lcao.test_dir.process+"time_"+ModuleBase::GlobalFunc::TO_STRING(GlobalV::MY_RANK),std::ofstream::app);
-timeval t_start;
-gettimeofday( &t_start, NULL);
 	ModuleBase::TITLE("Matrix_Orbs11","cal_overlap_matrix");
 	
 	std::map<size_t,std::map<size_t,std::map<size_t,std::map<size_t,RI::Tensor<Tdata>>>>> matrixes;
@@ -156,8 +152,6 @@ gettimeofday( &t_start, NULL);
 			}
 		}
 	}
-// ofs<<"TIME@Exx_Abfs::Matrix_Orbs11::cal_overlap_matrix\t"<<time_during(t_start)<<std::endl;
-// ofs.close();
 	return matrixes;
 }
